Partition member body facts fixtures
Move the constructor, namespace, and backing-field compiler fixtures beside their sole owning test class without changing declarations or method bodies.
